+++ NEWS.md
@@ -1,7 +1,10 @@
 # tbgeneratr 0.0.0.21.9003
 #### 13 Jan 2019
 ## Major changes
-* `aggregate_dst()` added to simplify `dst_baseliner()`
+* `aggregate_dst()` added to simplify `dst_baseliner()`.
+
+## Minor changes
+* `drug_baseliner()` args changed.
 
 ---